restructure ros package
