+++ src/icm20948_driver/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(truckbot)
+project(icm20948_driver)
 
 #Default to C++14
 if(NOT CMAKE_CXX_STANDARD)
@@ -12,20 +12,12 @@ endif()
 
 find_package(ament_cmake REQUIRED)
 find_package(rclcpp REQUIRED)
-find_package(std_msgs REQUIRED)
 find_package(sensor_msgs REQUIRED)
-find_package(nav_msgs REQUIRED)
 find_package(Eigen3 3.3 REQUIRED)
-find_package(fmt REQUIRED)
 
 
 INCLUDE_DIRECTORIES(
   include
-)
-  
-add_executable(motor_controller 
-  src/MotorControllerMiddleware.cpp
-  src/MotorController.cpp
 )
 
 add_executable(imu
@@ -33,23 +25,12 @@ add_executable(imu
   src/rpi5ICM20948.cpp
 )
 
-add_executable(ekf
-  src/EKFMiddleware.cpp
-  src/EKF.cpp
-)
-
-target_link_libraries(motor_controller lgpio)
-target_link_libraries(imu lgpio)
-target_link_libraries(ekf Eigen3::Eigen fmt::fmt)
-ament_target_dependencies(motor_controller rclcpp std_msgs sensor_msgs)
+target_link_libraries(imu lgpio Eigen3::Eigen)
 ament_target_dependencies(imu rclcpp sensor_msgs)
-ament_target_dependencies(ekf rclcpp sensor_msgs nav_msgs)
 
 
 install(TARGETS
-  motor_controller
   imu
-  ekf
   DESTINATION lib/${PROJECT_NAME})
 
 ament_package()
