+++ src/icm20948_driver/package.xml
@@ -1,16 +1,14 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>truckbot</name>
-  <version>0.0.0</version>
+  <name>icm20948_driver</name>
+  <version>0.1.0</version>
   <description>TODO: Package description</description>
   <maintainer email="emory@todo.todo">emory</maintainer>
   <license>TODO: License declaration</license>
   <depend>rclcpp</depend>
-  <depend>std_msgs</depend>
   <depend>sensor_msgs</depend>
-  <depend>nav_msgs</depend>
-
+  <depend>Eigen3</depend>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
